modal page for blood sugars
