+++ src/components/index.ts
@@ -9,3 +9,4 @@ export {BsInformation} from './bs-information'
 export {ContentLoadingSegment} from './content-loading-segment'
 export {Picker} from './picker'
 export {BpModal} from './bp-modal'
+export {BsModal} from './bs-modal'
