+++ src/constants/screens.ts
@@ -11,10 +11,8 @@ enum SCREENS {
   CONTACT_A_DOCTOR = 'CONTACT_A_DOCTOR',
   HOME = 'HOME',
   BP_HISTORY = 'BP_HISTORY',
-  BP_DETAILS = 'BP_DETAILS',
   ADD_BP = 'ADD_BP',
   BS_HISTORY = 'BS_HISTORY',
-  BS_DETAILS = 'BS_DETAILS',
   ADD_BS = 'ADD_BS',
   ADD_MEDICINE = 'ADD_MEDICINE',
   DETAILS_MODAL_SCREEN = 'DETAILS_MODAL_SCREEN',
